+++ packages/core/src/index.ts
@@ -62,5 +62,3 @@ export * from './model'
 export * from './textmodel'
 export * from './tx'
 export * from './domains'
-export * from './objectid'
-export * from './utils'
